+++ NWN.Anvil/src/main/API/Events/Game/ModuleEvents/ModuleEvents.OnUserDefined.cs
@@ -1,0 +1,42 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Global module events.
+  /// </summary>
+  public static partial class ModuleEvents
+  {
+    [GameEvent(EventScriptType.ModuleOnUserDefinedEvent)]
+    public sealed class OnUserDefined : IEvent
+    {
+      public int EventNumber { get; } = NWScript.GetUserDefinedEventNumber();
+
+      NwObject IEvent.Context
+      {
+        get => null;
+      }
+
+      public static void Signal(int eventId)
+      {
+        Event nwEvent = NWScript.EventUserDefined(eventId);
+        NWScript.SignalEvent(NwModule.Instance, nwEvent);
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwModule
+  {
+    /// <inheritdoc cref="ModuleEvents.OnUserDefined"/>
+    public event Action<ModuleEvents.OnUserDefined> OnUserDefined
+    {
+      add => EventService.SubscribeAll<ModuleEvents.OnUserDefined, GameEventFactory, GameEventFactory.RegistrationData>(new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.UnsubscribeAll<ModuleEvents.OnUserDefined, GameEventFactory>(value);
+    }
+  }
+}
